refactor: bring out components from slug/new into their own files
